requirements.txt
langchain
langchain_community
langchain-huggingface
langchain_mistralai
langchain_huggingface

python-dotenv

mistralai

streamlit

PyPDF

faiss-cpu

pinecone

langchain_pinecone